✨ Refactor navbar components for better organization.
