+++ app/_components/header.js
@@ -2,7 +2,7 @@ import Image from "next/image";
 import Link from "next/link";
 import bg from "public/images/home/bg.png";
 import React from "react";
-import Navbar from "./navbar";
+import Navbar from "./navbars";
 
 /**
  * Renders a social icon with a link.
